Fix build for macOS and Linux CI

+++ test/unit/ak_audit_test.c
@@ -113,10 +113,10 @@ static void test_sha256(const uint8_t *data, uint32_t len, uint8_t *output)
     }
 
     for (int i = 0; i < 8; i++) {
-        output[i*4] = (h[i] >> 24) & 0xFF;
-        output[i*4+1] = (h[i] >> 16) & 0xFF;
-        output[i*4+2] = (h[i] >> 8) & 0xFF;
-        output[i*4+3] = h[i] & 0xFF;
+        output[i*4] = (uint8_t)((h[i] >> 24) & 0xFF);
+        output[i*4+1] = (uint8_t)((h[i] >> 16) & 0xFF);
+        output[i*4+2] = (uint8_t)((h[i] >> 8) & 0xFF);
+        output[i*4+3] = (uint8_t)(h[i] & 0xFF);
     }
 }
